requirements.txt
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
mysql-connector-python==8.3.0
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
tkcalendar==1.6.1
urllib3==2.2.1
